Clean aliases examples (#318)
The scope of this PR is mostly to fix the aliases page to work with the
new testing framework, excluding the examples we cannot test.

1. make all syntax example blocks into `[source, syntax]`, and remove
the bulky header from those.
1. exclude remote aliases from testing. It is too complicated to set up
the needed infrastructure to test them, both server and client side. Not
a priority for now.
1. as a result of 2., change the setup queries so that our new testing
doesn't pick up the remote aliases commands.
1. add `WAIT` on database creation in test setup queries, so that we
don't get un-ready databases in subsequent queries.
1. exclude aliases with dots from testing, when the dots are not
separators between a composite db name and an alias name. This is
because it becomes impossible to automate their deletion by taking the
list of aliases from `SHOW ALIASES` and deleting them one by one, as the
escaping is lost.
1. remove all `indent=0` attributes in code blocks, which had no effect.

Co-authored-by: Jens Pryce-Åklundh <112686610+JPryce-Aklundh@users.noreply.github.com>