Make `Delete` operator example query plan more deterministic (#434)
The example for the `Delete` operator currently has a query plan that is
not guaranteed to be the same every time, even with the same indexes.
Given the presence of multiple indexes that can be used for the query,
the planner will try to figure out which gives best performance, but
with so little data in the database all plans perform roughly the same,
**so the decision over which plan is chosen is arbitrary**. The result
is that we sometimes get `NodeIndexSeek` as first operator, sometimes
`DirectedRelationshipIndexSeek` (which leverages the index on
`WORKS_IN.duration`).

This PR tweaks the example so that the ambiguity is removed, and only
one query plan can come up. All in all, it's also an improvement, since:
1. it still showcases the `Delete` operator, which is the whole point,
but shortens the list of operators and makes `Delete` more immediately
visible
2. it makes the example more uniform with respect to the following one
using `DETACH DELETE`, by using a more similar query.