Restructure index section (#500)
This will hopefully remove some confusion for users around just showing
off the command vs an example as well as removing examples for index
providers when only one provider exists.

It also updates the links to not mention `administration` as the index
chapter does not live in the administration chapter.