5.8 dev reconciliation (#564)
This PR updates 5.x with updates from PRs into dev that were not
cherry-picked into 5.x at any point.

While it looks like a lot of commits were missing, in the end it amounts
to changes in 14 files, some of which are trivial and do not affect the
content or output.

I'd be happy to use 'squash and merge' to reduce this update to a single
commit in 5.x