Remove Administration pages + Keyword Glossary page (#766)
The Database administration pages have now been migrated to the
Operations Manual.
This PR deletes the pages in the Cypher Manual, leaving only the
administration/index.adoc where a note is posted about content
relocation.
This PR also deletes the keyword glossary page, which was not maintained
and is just a poor substitute for the cheat cheet.

---------

Co-authored-by: Reneta Popova <reneta.popova@neo4j.com>