Add docs for the new nullIf() function (#767)
NullIf can be useful for handling special values like 0 or empty strings
by replacing them with null, it can also be used in combination with
coalesce to substitute the null values with something else. Another use
case is to avoid division by zero errors, this is because division by
null returns null, whereas division by 0 will error! It is a handy
shorthand that SQL users will be familiar with.

Co-authored-by: Jens Pryce-Åklundh <112686610+JPryce-Aklundh@users.noreply.github.com>