Fix some tests (#850)
- A setup block was (silently) failing due to an extra trailing comma.
With that fixed, the test for the other example failed because `Charlie`
was not reported as a string in the asciidoc.
- `test-result-skip` is not needed when a result is missing, only when a
result is present but _different_ from what the tester would get.
- No need to `test-skip` when query parameters are missing: the tester
skips automatically.
- After a query block, there may only be _one_ result block, either of
type `queryresult` or `queryplan`. If there's both of them, the parser
will pick the first one, but its internal state will probably also get
messed up.
- The last two query plans were surely wrong (they didn't have the right
return value), while I'm not sure about the first. It may be something
that changed in newer releases or IDK, they look pretty similar.

---------

Co-authored-by: Jens Pryce-Åklundh <112686610+JPryce-Aklundh@users.noreply.github.com>